update fixture production
